+++ src/js/gutenberg-overrides/edit-post/index.js
@@ -1,6 +1,2 @@
-import def from 'gutenberg/edit-post?source=node_modules';
-
 export { PluginDocumentSidebarPanel } from './components/sidebar/document-sidebar';
-
-export default def;
 export * from 'gutenberg/edit-post?source=node_modules';