+++ src/js/gutenberg-overrides/editor/store/actions.js
@@ -7,8 +7,7 @@ import {
 } from '@wordpress/blocks';
 
 import * as paragraph from 'gutenberg/core-blocks/paragraph';
-
-import def, * as others from 'gutenberg/editor/store/actions?source=node_modules';
+import * as others from 'gutenberg/editor/store/actions?source=node_modules';
 
 const { insertBlock } = others;
 
@@ -17,12 +16,11 @@ const { insertBlock } = others;
 others.insertDefaultBlock = (attributes, rootUID, index) => {
   const blockName = rootUID === undefined ? getDefaultBlockName() : paragraph.name;
   const block = createBlock(blockName, attributes);
-
+  
   return {
     ...insertBlock(block, index, rootUID),
-    isProvisional: blockName === paragraph.name,
+    isProvisional: true, // blockName === paragraph.name,
   };
 };
 
-export default def;
 export * from 'gutenberg/editor/store/actions?source=node_modules';