+++ src/js/gutenberg-overrides/core-blocks/CHANGELOG.md
@@ -4,4 +4,8 @@
 
 ### Added
 
-- `data` attribute to `cover-image` block (`cover-image/index.js`)
+- `data` attribute to `cover-image` block and change *onSelectImage* function to set `data` atributes (`cover-image/index.js`)
+
+- `data` attribute to `gallery` block and **GalleryImage** (`gallery/index.js`, `gallery/edit.js` and `gallery/gallery-image.js`)
+
+- `data` attribute to `image` block and change *onSelectImage* function to set `data` atributes (`image/index.js` and `image/edit.js`)